Move tests for object details
